garbage/pyproject.toml
[project]
name = "garbage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
]

[tool.uv.workspace]
members = [
    "project",
]
